v l@st

+++ ft_putadress.c
@@ -6,7 +6,7 @@
 /*   By: sasaidi <sasaidi@student.1337.ma>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/12 11:51:56 by sasaidi           #+#    #+#             */
-/*   Updated: 2025/11/12 22:26:16 by sasaidi          ###   ########.fr       */
+/*   Updated: 2025/11/13 21:42:51 by sasaidi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,11 +14,11 @@
 
 int	ft_putadress(void *ptr)
 {
-	int					count;
-	unsigned long long	adress;
+	int				count;
+	unsigned long	adress;
 
 	count = 0;
-	adress = (unsigned long long)ptr;
+	adress = (unsigned long)ptr;
 	count += ft_putstr("0x");
 	if (adress == 0)
 		count += ft_putchar('0');
