+++ ft_puthexa_for_adrs.c
@@ -6,13 +6,13 @@
 /*   By: sasaidi <sasaidi@student.1337.ma>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/12 20:36:58 by sasaidi           #+#    #+#             */
-/*   Updated: 2025/11/13 17:46:18 by sasaidi          ###   ########.fr       */
+/*   Updated: 2025/11/13 21:43:00 by sasaidi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libftprintf.h"
 
-int	ft_puthexa_long(unsigned long long nb)
+int	ft_puthexa_long(unsigned long nb)
 {
 	char	arr[16];
 	int		i;
